refactor: password page query 분리
