+++ src/views/PasswordPage/hooks/useMutateChangePassword.tsx
@@ -1,0 +1,26 @@
+import { useMutation } from '@tanstack/react-query';
+import { AxiosError, AxiosResponse } from 'axios';
+import { useNavigate } from 'react-router-dom';
+
+import { sendingPasswordChange } from '../apis';
+
+import type { PasswordError, PasswordRequest, PasswordResponse } from '../types';
+
+const useMutateChangePassword = () => {
+  const navigate = useNavigate();
+
+  const { mutate: changePasswordMutate, isPending: changePasswordIsPending } = useMutation<
+    AxiosResponse<PasswordResponse, PasswordRequest>,
+    AxiosError<PasswordError, PasswordRequest>,
+    PasswordRequest
+  >({
+    mutationFn: (userInfo: PasswordRequest) => sendingPasswordChange(userInfo),
+    onSuccess: () => {
+      navigate('/');
+    },
+  });
+
+  return { changePasswordMutate, changePasswordIsPending };
+};
+
+export default useMutateChangePassword;
